+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Delete
 - Remove Prophecy `phpspec/prophecy` usage for testing
+- Remove `defuse/php-encryption` support (only Sodium library used)
 
 ## [1.0.0] - 2020-10-12
 ### Added